+++ src/StorageBase/Action/PortalExtension/Find/PortalExtensionFindResult.php
@@ -6,6 +6,7 @@ namespace Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Find;
 
 use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
 use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringReferenceContract;
 use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalExtensionContract;
 
@@ -14,7 +15,7 @@ final class PortalExtensionFindResult implements AttachmentAwareInterface
     use AttachmentAwareTrait;
 
     /**
-     * @var array<class-string<PortalExtensionContract>, bool>
+     * @var array<string, bool>
      */
     private array $extensions = [];
 
@@ -23,16 +24,13 @@ final class PortalExtensionFindResult implements AttachmentAwareInterface
         $this->attachments = new AttachmentCollection();
     }
 
-    /**
-     * @param class-string<PortalExtensionContract> $class
-     */
-    public function add(string $class, bool $active): void
+    public function add(ClassStringReferenceContract $portalExtensionType, bool $active): void
     {
-        $this->extensions[$class] ??= $active;
+        $this->extensions[(string) $portalExtensionType] ??= $active;
     }
 
     public function isActive(PortalExtensionContract $portalExtension): bool
     {
-        return $this->extensions[\get_class($portalExtension)] ?? $portalExtension->isActiveByDefault();
+        return $this->extensions[(string) $portalExtension::class()] ?? $portalExtension->isActiveByDefault();
     }
 }