+++ src/StorageBase/MappingCollection.php
@@ -38,8 +38,8 @@ class MappingCollection extends AbstractObjectCollection
         foreach ($this->items as $mapping) {
             $entityType = $mapping->getEntityType();
 
-            $typedMappings[$entityType] ??= new TypedMappingCollection($entityType);
-            $typedMappings[$entityType]->push([$mapping]);
+            $typedMappings[(string) $entityType] ??= new TypedMappingCollection($entityType);
+            $typedMappings[(string) $entityType]->push([$mapping]);
         }
 
         yield from $typedMappings;