+++ src/UiAdminBase/Action/PortalNode/PortalNodeAdd/PortalNodeAddResult.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeAdd;
+
+use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
+
+final class PortalNodeAddResult implements AttachmentAwareInterface
+{
+    use AttachmentAwareTrait;
+
+    private PortalNodeKeyInterface $portalNodeKey;
+
+    public function __construct(PortalNodeKeyInterface $portalNodeKey)
+    {
+        $this->attachments = new AttachmentCollection();
+        $this->portalNodeKey = $portalNodeKey;
+    }
+
+    public function getPortalNodeKey(): PortalNodeKeyInterface
+    {
+        return $this->portalNodeKey;
+    }
+}