+++ src/UiAdminBase/Contract/Action/PortalNode/PortalNodeExtensionActivateUiActionInterface.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\PortalNode;
+
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionActivate\PortalNodeExtensionActivatePayload;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\InvalidArgumentThrowableInterface;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\NoMatchForPackageQueryException;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalExtensionsAreAlreadyActiveOnPortalNodeException;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalNodeMissingException;
+
+interface PortalNodeExtensionActivateUiActionInterface
+{
+    /**
+     * Activate portal extensions by the given queries on the given portal node.
+     *
+     * @throws InvalidArgumentThrowableInterface
+     * @throws NoMatchForPackageQueryException
+     * @throws PortalExtensionsAreAlreadyActiveOnPortalNodeException
+     * @throws PortalNodeMissingException
+     */
+    public function activate(PortalNodeExtensionActivatePayload $payload): void;
+}