+++ src/UiAdminBase/Contract/Exception/UiAdminBaseThrowableInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception;
+
+/**
+ * Identifies every exception from this package
+ */
+interface UiAdminBaseThrowableInterface extends \Throwable
+{
+}