+++ test/PortalBase/Builder/FlowComponentTest.php
@@ -43,6 +43,16 @@ use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerInterface;
 
 /**
+ * @covers \Heptacom\HeptaConnect\Core\Portal\PortalConfiguration
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringReferenceContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\SubtypeClassStringContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\DatasetEntityCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\EntityType
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\TypedDatasetEntityCollection
  * @covers \Heptacom\HeptaConnect\Portal\Base\Builder\Builder\HttpHandlerBuilder
  * @covers \Heptacom\HeptaConnect\Portal\Base\Builder\Component\Emitter
  * @covers \Heptacom\HeptaConnect\Portal\Base\Builder\Component\Explorer
@@ -60,6 +70,14 @@ use Psr\Log\LoggerInterface;
  * @covers \Heptacom\HeptaConnect\Portal\Base\Builder\FlowComponent
  * @covers \Heptacom\HeptaConnect\Portal\Base\Builder\Token\HttpHandlerToken
  * @covers \Heptacom\HeptaConnect\Portal\Base\Builder\ResolveArgumentsTrait
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Emission\Contract\EmitterContract
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Emission\EmitterStack
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExplorerContract
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Exploration\ExplorerStack
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiverContract
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Reception\ReceiverStack
+ * @covers \Heptacom\HeptaConnect\Portal\Base\StatusReporting\Contract\StatusReporterContract
+ * @covers \Heptacom\HeptaConnect\Portal\Base\StatusReporting\StatusReporterStack
  * @covers \Heptacom\HeptaConnect\Portal\Base\Web\Http\Contract\HttpHandlerContract
  * @covers \Heptacom\HeptaConnect\Portal\Base\Web\Http\HttpHandlerStack
  */
@@ -418,31 +436,31 @@ final class FlowComponentTest extends TestCase
         $thisClasses = [];
         $supports = [];
 
-        $explorerToken = new ExplorerToken(FirstEntity::class);
+        $explorerToken = new ExplorerToken(FirstEntity::class());
         $explorerToken->setRun(function () use (&$supports, &$thisClasses): array {
             /* @var $this ExplorerContract */
             $thisClasses[] = static::class;
-            $supports[] = $this->supports();
+            $supports[] = (string) $this->getSupportedEntityType();
 
             return [];
         });
         $explorer = new Explorer($explorerToken);
 
-        $emitterToken = new EmitterToken(FirstEntity::class);
+        $emitterToken = new EmitterToken(FirstEntity::class());
         $emitterToken->setBatch(function () use (&$supports, &$thisClasses): array {
             /* @var $this EmitterContract */
             $thisClasses[] = static::class;
-            $supports[] = $this->supports();
+            $supports[] = (string) $this->getSupportedEntityType();
 
             return [];
         });
         $emitter = new Emitter($emitterToken);
 
-        $receiverToken = new ReceiverToken(FirstEntity::class);
+        $receiverToken = new ReceiverToken(FirstEntity::class());
         $receiverToken->setBatch(function () use (&$supports, &$thisClasses): void {
             /* @var $this ReceiverContract */
             $thisClasses[] = static::class;
-            $supports[] = $this->supports();
+            $supports[] = (string) $this->getSupportedEntityType();
         });
         $receiver = new Receiver($receiverToken);
 
@@ -467,9 +485,9 @@ final class FlowComponentTest extends TestCase
         $httpHandler = new HttpHandler($httpHandlerToken);
 
         \iterable_to_array($explorer->explore($exploreContext, new ExplorerStack([])));
-        \iterable_to_array($emitter->emit([], $emitContext, new EmitterStack([], FirstEntity::class)));
+        \iterable_to_array($emitter->emit([], $emitContext, new EmitterStack([], FirstEntity::class())));
         \iterable_to_array($receiver->receive(
-            new TypedDatasetEntityCollection(FirstEntity::class, [new FirstEntity()]),
+            new TypedDatasetEntityCollection(FirstEntity::class(), [new FirstEntity()]),
             $receiveContext,
             new ReceiverStack([])
         ));