+++ test/UiAdminBase/Action/UiActionParameterTest.php
@@ -1,0 +1,90 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Ui\Admin\Base\Test\Action;
+
+use Heptacom\HeptaConnect\Core\Test\Fixture\FooBarEmitter;
+use Heptacom\HeptaConnect\Core\Test\Fixture\FooBarPortalExtension;
+use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\UnsafeClassString;
+use Heptacom\HeptaConnect\Portal\Base\FlowComponent\CodeOrigin;
+use Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey;
+use Heptacom\HeptaConnect\Storage\Base\Test\Fixture\FirstEntity;
+use Heptacom\HeptaConnect\Storage\Base\Test\Fixture\Portal;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListCriteria;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListResult;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeAdd\PortalNodeAddPayload;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeAdd\PortalNodeAddResult;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionActivate\PortalNodeExtensionActivatePayload;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseCriteria;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseResult;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionDeactivate\PortalNodeExtensionDeactivatePayload;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeStatusReport\PortalNodeStatusReportPayload;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeStatusReport\PortalNodeStatusReportResult;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringReferenceContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\SubtypeClassStringContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\EntityType
+ * @covers \Heptacom\HeptaConnect\Portal\Base\FlowComponent\CodeOrigin
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalExtensionContract
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionType
+ * @covers \Heptacom\HeptaConnect\Portal\Base\Portal\PortalType
+ * @covers \Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListCriteria
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListResult
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeAdd\PortalNodeAddPayload
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeAdd\PortalNodeAddResult
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionActivate\PortalNodeExtensionActivatePayload
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseCriteria
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionBrowse\PortalNodeExtensionBrowseResult
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeExtensionDeactivate\PortalNodeExtensionDeactivatePayload
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeStatusReport\PortalNodeStatusReportPayload
+ * @covers \Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeStatusReport\PortalNodeStatusReportResult
+ */
+final class UiActionParameterTest extends TestCase
+{
+    public function testAttachabilityOfStorageActionStructs(): void
+    {
+        foreach ($this->iterateAttachmentAwareActionStructs() as $attachmentAware) {
+            $attachment = new FirstEntity();
+            $attachmentAware->attach($attachment);
+            static::assertTrue($attachmentAware->isAttached($attachment));
+            $attachmentAware->detach($attachment);
+        }
+    }
+
+    /**
+     * @return iterable<AttachmentAwareInterface>
+     */
+    private function iterateAttachmentAwareActionStructs(): iterable
+    {
+        $portalClass = Portal::class;
+        $entityType = FirstEntity::class;
+        $codeOrigin = new CodeOrigin(__FILE__, 0, 1);
+        $portalNodeKey = new PreviewPortalNodeKey($portalClass::class());
+        $portalExtensionClass = FooBarPortalExtension::class;
+        $unsafeClass = new UnsafeClassString($entityType);
+
+        yield new PortalEntityListCriteria($portalClass::class());
+        yield new PortalEntityListResult($codeOrigin, $entityType::class(), FooBarEmitter::class);
+        yield new PortalEntityListResult($codeOrigin, $unsafeClass, FooBarEmitter::class);
+        yield new PortalNodeAddPayload($portalClass::class());
+        yield new PortalNodeAddResult($portalNodeKey);
+        yield new PortalNodeExtensionActivatePayload($portalNodeKey);
+        yield new PortalNodeExtensionBrowseCriteria($portalNodeKey);
+        yield new PortalNodeExtensionBrowseResult($portalNodeKey, true, $portalExtensionClass::class());
+        yield new PortalNodeExtensionBrowseResult($portalNodeKey, true, $unsafeClass);
+        yield new PortalNodeExtensionDeactivatePayload($portalNodeKey);
+        yield new PortalNodeStatusReportPayload($portalNodeKey, []);
+        yield new PortalNodeStatusReportResult($portalNodeKey, '', true, []);
+    }
+}